Works!
My understanding is that it should report scrolling within an element. Say I have a div that has content too long for it's height and it displays a scrollbar. The page itself may be long and have scrollbars. When I attach a listener to the div, I expect it to be fired when I scroll the contents in the div and not when I scroll the page. For page scroll, I would attach a listener on the page. It still doesn't work for the document element. Will fix it if needed.